+++ lib/pubsub/pub_sub_config.d.ts
@@ -1,6 +1,7 @@
 export interface RfiPubSubConfig {
     keyFilename: string;
     topicPrefix: string;
+    resetHostedSubscriptions: boolean;
 }
 export declare const PubSubConfig: {
     keyFilename: {