+++ src/server/rumbleship-context.ts
@@ -12,6 +12,8 @@ export interface RumbleshipContextOptionsPlain {
   logger?: SpyglassLogger;
   container?: ContainerInstance;
   initial_trace_metadata?: object;
+  marshalled_trace?: string;
+  linked_span?: HoneycombSpan;
 }
 
 class RumbleshipContextOptionsWithDefaults {
@@ -21,6 +23,8 @@ class RumbleshipContextOptionsWithDefaults {
   private readonly _id: string;
   private readonly _initial_trace_metadata: object;
   private readonly _config: object;
+  private readonly _marshalled_trace?: string;
+  private readonly _linked_span?: HoneycombSpan;
   get authorizer() {
     return this._authorizer;
   }
@@ -39,9 +43,17 @@ class RumbleshipContextOptionsWithDefaults {
   get config() {
     return this._config;
   }
+  get marshalled_trace() {
+    return this._marshalled_trace;
+  }
+  get linked_span() {
+    return this._linked_span;
+  }
   constructor(filename: string, options: RumbleshipContextOptionsPlain) {
     this._config = options.config;
     this._id = options.id ?? uuid.v4();
+    this._marshalled_trace = options.marshalled_trace;
+    this._linked_span = options.linked_span;
     this._initial_trace_metadata = options.initial_trace_metadata
       ? { name: 'app.rumbleship_context', ...options.initial_trace_metadata }
       : {
@@ -106,10 +118,15 @@ export class RumbleshipContext implements Context {
     if (!this.initialized) {
       throw new Error('Must initialize the RumbleshipContext before making instances');
     }
-    const { authorizer, container, id, logger } = new RumbleshipContextOptionsWithDefaults(
-      filename,
-      options
-    );
+    const {
+      authorizer,
+      container,
+      id,
+      logger,
+      initial_trace_metadata,
+      marshalled_trace,
+      linked_span
+    } = new RumbleshipContextOptionsWithDefaults(filename, options);
 
     container.set('logger', logger);
     container.set('authorizer', authorizer);
@@ -119,7 +136,16 @@ export class RumbleshipContext implements Context {
     }
 
     const beeline = container.get<typeof RumbleshipBeeline>('beelineFactory').make(id);
-    const ctx = new RumbleshipContext(id, container, logger, authorizer, beeline);
+    const ctx = new RumbleshipContext(
+      id,
+      container,
+      logger,
+      authorizer,
+      beeline,
+      initial_trace_metadata,
+      marshalled_trace,
+      linked_span
+    );
     RumbleshipContext.ActiveContexts.set(ctx.id, ctx);
     logger.debug(`NEW SERVICE CONTEXT: ${ctx.id}`);
     const withSequelize = this.addSequelizeServicesToContext(ctx) as RumbleshipContext;
@@ -131,17 +157,32 @@ export class RumbleshipContext implements Context {
     public container: ContainerInstance,
     public logger: SpyglassLogger,
     public authorizer: Authorizer,
-    public beeline: RumbleshipBeeline
-  ) {}
+    public beeline: RumbleshipBeeline,
+    initial_trace_metadata: object,
+    marshalled_trace?: string,
+    linked_span?: HoneycombSpan
+  ) {
+    const hydrated_trace = this.beeline.unmarshalTraceContext(marshalled_trace) as HoneycombSpan;
+    this.trace = this.beeline.startTrace(
+      { name: 'rumbleship_context', ...initial_trace_metadata },
+      hydrated_trace.traceId,
+      hydrated_trace.parentSpanId,
+      hydrated_trace.dataset
+    );
+    if (linked_span) {
+      this.beeline.linkToSpan(linked_span!);
+    }
+  }
 
   release() {
     if (this.trace) {
-      this.beeline.finishRumbleshipContextTrace();
+      this.beeline.finishTrace(this.trace);
     }
     this.logger.debug(`RELEASE SERVICE CONTEXT: ${this.id}`);
     this.container.reset();
   }
 }
+/** @deprecated ? */
 export function withRumbleshipContext<T>(
   filename: string,
   options: RumbleshipContextOptionsPlain,
@@ -157,7 +198,7 @@ export function withRumbleshipContext<T>(
   );
 
   return new Promise((resolve, reject) => {
-    const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx));
+    const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx))();
     if (isPromise(value)) {
       // tslint:disable-next-line: no-floating-promises
       ((value as unknown) as Promise<T>)
@@ -171,6 +212,7 @@ export function withRumbleshipContext<T>(
   });
 }
 
+/** @deprecated ? */
 export function withLinkedRumbleshipContext<T>(
   parentSpan: HoneycombSpan,
   filename: string,
@@ -190,7 +232,7 @@ export function withLinkedRumbleshipContext<T>(
     ctx.beeline.linkToSpan(parentSpan);
 
     return new Promise((resolve, reject) => {
-      const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx));
+      const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx))();
       if (isPromise(value)) {
         // tslint:disable-next-line: no-floating-promises
         ((value as unknown) as Promise<T>)
@@ -224,8 +266,10 @@ export function getRumbleshipContext(filename: string, config: object): Rumblesh
   return Factory.make(filename, { config });
 }
 
-export function releaseRumbleshipContext(context: Context) {
-  context.beeline.finishRumbleshipContextTrace();
+export function releaseRumbleshipContext(context: RumbleshipContext) {
+  if (context.trace) {
+    context.beeline.finishTrace(context.trace);
+  }
   context.logger.debug(`RELEASE SERVICE CONTEXT: ${context.id}`);
   context.container.reset();
 }