+++ src/server/rumbleship-context.ts
@@ -12,6 +12,8 @@ export interface RumbleshipContextOptionsPlain {
   logger?: SpyglassLogger;
   container?: ContainerInstance;
   initial_trace_metadata?: object;
+  marshalled_trace?: string;
+  linked_span?: HoneycombSpan;
 }
 
 class RumbleshipContextOptionsWithDefaults {
@@ -21,6 +23,8 @@ class RumbleshipContextOptionsWithDefaults {
   private readonly _id: string;
   private readonly _initial_trace_metadata: object;
   private readonly _config: object;
+  private readonly _marshalled_trace?: string;
+  private readonly _linked_span?: HoneycombSpan;
   get authorizer() {
     return this._authorizer;
   }
@@ -39,9 +43,17 @@ class RumbleshipContextOptionsWithDefaults {
   get config() {
     return this._config;
   }
+  get marshalled_trace() {
+    return this._marshalled_trace;
+  }
+  get linked_span() {
+    return this._linked_span;
+  }
   constructor(filename: string, options: RumbleshipContextOptionsPlain) {
     this._config = options.config;
     this._id = options.id ?? uuid.v4();
+    this._marshalled_trace = options.marshalled_trace;
+    this._linked_span = options.linked_span;
     this._initial_trace_metadata = options.initial_trace_metadata
       ? { name: 'app.rumbleship_context', ...options.initial_trace_metadata }
       : {
@@ -106,10 +118,15 @@ export class RumbleshipContext implements Context {
     if (!this.initialized) {
       throw new Error('Must initialize the RumbleshipContext before making instances');
     }
-    const { authorizer, container, id, logger } = new RumbleshipContextOptionsWithDefaults(
-      filename,
-      options
-    );
+    const {
+      authorizer,
+      container,
+      id,
+      logger,
+      initial_trace_metadata,
+      marshalled_trace,
+      linked_span
+    } = new RumbleshipContextOptionsWithDefaults(filename, options);
 
     container.set('logger', logger);
     container.set('authorizer', authorizer);
@@ -119,7 +136,16 @@ export class RumbleshipContext implements Context {
     }
 
     const beeline = container.get<typeof RumbleshipBeeline>('beelineFactory').make(id);
-    const ctx = new RumbleshipContext(id, container, logger, authorizer, beeline);
+    const ctx = new RumbleshipContext(
+      id,
+      container,
+      logger,
+      authorizer,
+      beeline,
+      initial_trace_metadata,
+      marshalled_trace,
+      linked_span
+    );
     RumbleshipContext.ActiveContexts.set(ctx.id, ctx);
     logger.debug(`NEW SERVICE CONTEXT: ${ctx.id}`);
     const withSequelize = this.addSequelizeServicesToContext(ctx) as RumbleshipContext;
@@ -131,17 +157,32 @@ export class RumbleshipContext implements Context {
     public container: ContainerInstance,
     public logger: SpyglassLogger,
     public authorizer: Authorizer,
-    public beeline: RumbleshipBeeline
-  ) {}
+    public beeline: RumbleshipBeeline,
+    initial_trace_metadata: object,
+    marshalled_trace?: string,
+    linked_span?: HoneycombSpan
+  ) {
+    const hydrated_trace = this.beeline.unmarshalTraceContext(marshalled_trace) as HoneycombSpan;
+    this.trace = this.beeline.startTrace(
+      { name: 'rumbleship_context', ...initial_trace_metadata },
+      hydrated_trace.traceId,
+      hydrated_trace.parentSpanId,
+      hydrated_trace.dataset
+    );
+    if (linked_span) {
+      this.beeline.linkToSpan(linked_span!);
+    }
+  }
 
   release() {
     if (this.trace) {
-      this.beeline.finishRumbleshipContextTrace();
+      this.beeline.finishTrace(this.trace);
     }
     this.logger.debug(`RELEASE SERVICE CONTEXT: ${this.id}`);
     this.container.reset();
   }
 }
+/** @deprecated ? */
 export function withRumbleshipContext<T>(
   filename: string,
   options: RumbleshipContextOptionsPlain,
@@ -157,7 +198,7 @@ export function withRumbleshipContext<T>(
   );
 
   return new Promise((resolve, reject) => {
-    const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx));
+    const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx))();
     if (isPromise(value)) {
       // tslint:disable-next-line: no-floating-promises
       ((value as unknown) as Promise<T>)
@@ -171,63 +212,8 @@ export function withRumbleshipContext<T>(
   });
 }
 
-export function withLinkedRumbleshipContext<T>(
-  parentSpan: HoneycombSpan,
-  filename: string,
-  options: RumbleshipContextOptionsPlain,
-  fn: (ctx: RumbleshipContext) => T
-): Promise<T> {
-  const { initial_trace_metadata } = new RumbleshipContextOptionsWithDefaults(filename, options);
-  const ctx = Container.get<typeof RumbleshipContext>('RumbleshipContext').make(filename, options);
-  return ctx.beeline.runWithoutTrace(() => {
-    ctx.trace = ctx.beeline.startTrace(
-      {
-        ...initial_trace_metadata
-      },
-      ctx.id
-    );
-
-    ctx.beeline.linkToSpan(parentSpan);
-
-    return new Promise((resolve, reject) => {
-      const value = ctx.beeline.bindFunctionToTrace(() => fn(ctx));
-      if (isPromise(value)) {
-        // tslint:disable-next-line: no-floating-promises
-        ((value as unknown) as Promise<T>)
-          .then(resolve)
-          .catch(reject)
-          .finally(() => ctx.release());
-      } else {
-        ctx.release();
-        resolve(value);
-      }
-    });
-  });
-}
-
 function isPromise(p: any) {
   return p && typeof p.then === 'function';
-}
-
-/**
- * Provides a context that has an authorizer and credentials etc specifically for
- * THIS microservice so it can be used outside of the context of an Http/geaphql request
- * or GQL subscription.
- */
-export function getRumbleshipContext(filename: string, config: object): RumbleshipContext {
-  // tslint:disable-next-line: no-console
-  console.warn(`You probably DO NOT WANT to use this. 
-  \tYou probably want to wrap your code with \`RumbleshipContext.withContext(cb: () => any)\`
-  \tso you don't have to manage releasing the created context.
-  \tIf you meant to use this, YOU MUST INVOKE \`context.release()\` otherwise memory leaks!`);
-  const Factory = Container.get('RumbleshipContext') as typeof RumbleshipContext;
-  return Factory.make(filename, { config });
-}
-
-export function releaseRumbleshipContext(context: Context) {
-  context.beeline.finishRumbleshipContextTrace();
-  context.logger.debug(`RELEASE SERVICE CONTEXT: ${context.id}`);
-  context.container.reset();
 }
 
 export interface SpyglassLogger {