+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+### Removed
+### Changed
+  * Modified RumbleshipContext to only finish a trace if there is a trace set in the context.
+  * added resetHostedSubscriptions to the RfiPubSubConfig interface
+### Fixed
+### Deprecated
+### Security
+
+
 ## [6.0.2] -- 2020-04-20
 
 ### Changed