+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
   * RumbleshipContext.make() takes an optional `marshalled_trace` to propagate, optional `linked_span` to link to
-  * **IMPORTANT** Peer dependency on `@rumbleship/o11y` bumped to 2.0.1
+  * **IMPORTANT** Peer dependency on `@rumbleship/o11y` bumped to 3.0.0
 ### Removed
 ### Changed
   * Instantiating a RumbleshipContext automatically starts its (possibly distributed, possible linked) trace.